feat(backend): retry-opt-for-reverse-post #11650
