+++ dbm-services/mysql/db-tools/mysql-monitor/pkg/monitoriteminterface/internal/proxy.go
@@ -1,9 +1,9 @@
 package internal
 
 import (
-	"dbm-services/common/reverseapi"
 	reversemysqlapi "dbm-services/common/reverseapi/apis/mysql"
 	reversemysqldef "dbm-services/common/reverseapi/define/mysql"
+	"dbm-services/common/reverseapi/pkg/core"
 
 	"dbm-services/mysql/db-tools/dbactuator/pkg/core/cst"
 	"dbm-services/mysql/db-tools/dbactuator/pkg/util"
@@ -219,7 +219,7 @@ func addExLock() (fl *flock.Flock, err error) {
 }
 
 func confirmSelfIsRunning() (bool, error) {
-	apiCore, err := reverseapi.NewCore(int64(*config.MonitorConfig.BkCloudID))
+	apiCore, err := core.NewCore(int64(*config.MonitorConfig.BkCloudID), core.DefaultRetryOpts...)
 	if err != nil {
 		slog.Error("new core api error", slog.String("err", err.Error()))
 		return false, err
